bot_telegram_manicure/core/config.py

import os
from dotenv import load_dotenv
dotenv_path = os.path.join(os.path.dirname(__file__), '.env')
if os.path.exists(dotenv_path):
    load_dotenv(dotenv_path)





API_BOT1_KEY = os.environ.get("API_BOT1_KEY", None)
API_BOT2_KEY = os.environ.get("API_BOT2_KEY", None)

DB_NAME=os.getenv('DB_NAME')
DB_USER=os.getenv('DB_USER')
DB_PASSWORD=os.getenv('DB_PASSWORD')
DB_HOST=os.getenv('DB_HOST')